[FIX] user 구조변경
