+++ components/modal/joinedAgenda/index.js
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState, useRef } from "react";
 import Modal from "./agendaCard";
 import Card from "../../../ui/Card/Card";
 import { loadGetInitialProps } from "next/dist/shared/lib/utils";
@@ -6,8 +6,41 @@ import Image from "next/image";
 import Images from "../../../public/joined.png";
 import styles from "./JoinedAgenda.module.css";
 import { MdOutlineHowToVote } from "react-icons/md";
-const JoinedAgenda = ({ joinedAgenda }) => {
+import {
+  collection,
+  doc,
+  getDoc,
+  getFirestore,
+  limit,
+  onSnapshot,
+  orderBy,
+  query,
+  setDoc,
+  updateDoc,
+} from "firebase/firestore";
+const JoinedAgenda = ({ user }) => {
+  const [joinedAgenda, setJoinedAgenda] = useState([]);
+  const joinedAgendaUnsubsribe = useRef([]);
   const [showModal, setShowModal] = useState(false);
+  useEffect(async () => {
+    const db = getFirestore();
+    const joinedAgendaRef = collection(db, "user", user.uid, "joinedAgenda");
+    const joinedAgendaQuery = query(
+      joinedAgendaRef,
+      orderBy("joined"),
+      limit(10)
+    );
+    joinedAgendaUnsubsribe.current = await onSnapshot(
+      joinedAgendaQuery,
+      (snapshot) => {
+        const { length } = snapshot.docs;
+        console.log(length);
+        if (length > 0) {
+          setJoinedAgenda(snapshot.docs.map((str) => str.data()));
+        }
+      }
+    );
+  }, []);
   return (
     <div>
       <div className={styles.out} onClick={() => setShowModal(true)}>
