+++ styles/Modal.module.css
@@ -5,6 +5,7 @@
     border-radius: 15px;
     padding: 20px;
     z-index: 100;
+    overflow-y: scroll;
 }
 
 .body{
